+++ config/php-wake-on-lan.php
@@ -1,0 +1,11 @@
+<?php
+
+/*
+ * Package configuration settings.
+ */
+return [
+
+    'broadcast_address' => '255.255.255.255',
+
+    'route' => '/wake-on-lan/{macAddress}',
+];