+++ resources/views/result.blade.php
@@ -1,0 +1,11 @@
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <title>PHPWakeOnLan</title>
+    <meta name="description" content="PHPWakeOnLan">
+</head>
+<body>
+<pre>{{ print_r($result, true) }}</pre>
+</body>
+</html>