+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased]
+- Updated config files
+- Laravel support
+  - Config file
+  - Service provider
+  - Facade
+  - Console command
+  - View template
+  - Package route
+
+### Added
+- Initial Laravel support
+
 ### Changed
+- Main and tests class names
+- Exception error codes
+- Added PHP 7.1 support
+- Lowered phpunit minimum version
+- Updated README
+
+### Removed
+- Non compliant phpunit config flags
 
 ## [1.0.0] - 2019-02-28
 Initial release